pyproject.toml
[project]
name = "chatbot-using-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.3",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "langchain-unstructured>=0.1.5",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langsmith>=0.4.38",
    "newsdataapi>=0.1.27",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=0.2.5.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
